+++ scripts/app.js
@@ -202,6 +202,39 @@ function validateUrl(value, protocols = 'https?|ftp')
 	return new RegExp('^(?:(?:(?:'+protocols+'):)?\\/\\\/)(?:\\S+(?::\\S*)?@)?(?:(?!(?:10|127)(?:\\.\\d{1,3}){3})(?!(?:169\\.254|192\\.168)(?:\\.\\d{1,3}){2})(?!172\\.(?:1[6-9]|2\\d|3[0-1])(?:\\.\\d{1,3}){2})(?:[1-9]\\d?|1\\d\\d|2[01]\\d|22[0-3])(?:\\.(?:1?\\d{1,2}|2[0-4]\\d|25[0-5])){2}(?:\\.(?:[1-9]\\d?|1\\d\\d|2[0-4]\\d|25[0-4]))|(?:(?:[a-z\\u00a1-\\uffff0-9]-*)*[a-z\\u00a1-\\uffff0-9]+)(?:\.(?:[a-z\\u00a1-\\uffff0-9]-*)*[a-z\\u00a1-\\uffff0-9]+)*(?:\\.(?:[a-z\\u00a1-\\uffff]{2,})))(?::\\d{2,5})?(?:[/?#]\\S*)?$', 'i').test(value);
 }
 
+function copy(toCopy)
+{
+	if(typeof toCopy !== 'object' || toCopy === null)
+	{
+		return toCopy;
+	}
+	else if(Array.isArray(toCopy))
+	{
+		let len = toCopy.length;
+		let result = new Array(len);
+
+		for(let i = 0; i < len; i++)
+		{
+			let _toCopy = toCopy[i];
+			result[i] = typeof _toCopy !== 'object' || _toCopy === null ? _toCopy : copy(_toCopy);
+		}
+
+		return result;
+	}
+	else
+	{
+		let result = {};
+
+		for(let i in toCopy)
+		{
+			let _toCopy = toCopy[i];
+			result[i] = typeof _toCopy !== 'object' || _toCopy === null ? _toCopy : copy(_toCopy);
+		}
+
+		return result;
+	}
+}
+
 function time()
 {
 	return Math.floor(Date.now() / 1000);
@@ -242,6 +275,7 @@ module.exports = {
 	capitalize: capitalize,
 	stripTagsWithDOM: stripTagsWithDOM,
 	validateUrl: validateUrl,
+	copy: copy,
 	time: time,
 	sleep: sleep,
 	setImmediate: setImmediate,