+++ scripts/cache.js
@@ -21,7 +21,15 @@ function processTheImageQueue()
 	let realPath = fileManager.realPath(img.file);
 	let toImage = p.join(cacheFolder, sha+'.jpg');
 
-	image.resize(realPath, toImage, {width: img.size, height: Math.round(img.size * (img.poster ? 1.51369 : 1.5)), quality: 95, fit: img.poster ? 'cover' : 'inside'}).then(function(){
+	let fit = {
+		poster: 'cover',
+	};
+
+	let ratios = {
+		poster: 1.51369,
+	};
+
+	image.resize(realPath, toImage, {width: img.size, height: Math.round(img.size * (img.type ? ratios[img.type] : 1.5)), quality: 95, fit: img.type ? fit[img.type] : 'inside'}).then(function(){
 
 		if(typeof data[sha] == 'undefined') data[sha] = {lastAccess: app.time()};
 
@@ -66,9 +74,9 @@ function processTheImageQueue()
 	});
 }
 
-function addImageToQueue(file, size, sha, callback, vars, poster)
+function addImageToQueue(file, size, sha, callback, vars, type)
 {
-	queuedImages.push({file: file, size: size, sha: sha, callback: callback, vars: vars, poster: poster});
+	queuedImages.push({file: file, size: size, sha: sha, callback: callback, vars: vars, type: type});
 
 	if(!processingTheImageQueue && !stopTheImageQueue)
 	{
@@ -125,7 +133,10 @@ function returnThumbnailsImages(images, callback, file = false)
 	images = single ? [images] : images;
 
 	let size = Math.round(window.devicePixelRatio * 150);
-	let sizePoster = Math.round(window.devicePixelRatio * 146);
+
+	sizes = {
+		poster: Math.round(window.devicePixelRatio * 146),
+	};
 
 	let thumbnail = {};
 	let thumbnails = {};
@@ -138,17 +149,17 @@ function returnThumbnailsImages(images, callback, file = false)
 	{
 		let image = images[i];
 
-		let sha = (image.poster) ? sha1(image.path+'?type=poster') : (image.sha || sha1(image.path));
+		let sha = (image.type) ? sha1(image.path+'?type='+image.type) : (image.sha || sha1(image.path));
 		let imgCache = data[sha];
 
-		let _size = image.poster ? sizePoster : size;
+		let _size = image.type ? sizes[image.type] : size;
 
 		let path = addCacheVars(p.join(cacheFolder, sha+'.jpg'), _size, sha);
 
 		if(typeof imgCache == 'undefined' || !fs.existsSync(p.join(cacheFolder, sha+'.jpg')))
 		{
 			toGenerateThumbnails.push(image);
-			toGenerateThumbnailsData[image.path] = {sha: sha, vars: image.vars, poster: image.poster || false};
+			toGenerateThumbnailsData[image.path] = {sha: sha, vars: image.vars, type: image.type || false};
 
 			thumbnails[sha] = thumbnail = {cache: false, path: '', sha: sha};
 		}
@@ -159,7 +170,7 @@ function returnThumbnailsImages(images, callback, file = false)
 			if(imgCache.size != _size)
 			{
 				toGenerateThumbnails.push(image);
-				toGenerateThumbnailsData[image.path] = {sha: sha, vars: image.vars, poster: image.poster || false};
+				toGenerateThumbnailsData[image.path] = {sha: sha, vars: image.vars, type: image.type || false};
 
 				thumbnails[sha] = thumbnail = {cache: true, path: escapeBackSlash(path), sha: sha};
 			}
@@ -176,10 +187,10 @@ function returnThumbnailsImages(images, callback, file = false)
 		file.makeAvailable(toGenerateThumbnails, function(image) {
 
 			let data = toGenerateThumbnailsData[image.path];
-			let _size = data.poster ? sizePoster : size;
-			addImageToQueue(image.path, _size, data.sha, callback, data.vars || false, data.poster);
+			let _size = data.type ? sizes[data.type] : size;
+			addImageToQueue(image.path, _size, data.sha, callback, data.vars || false, data.type);
 
-		});
+		}, false, true);
 	}
 
 	return single ? thumbnail : thumbnails;
@@ -258,7 +269,7 @@ function readJsonInMemory(name)
 
 		jsonMemory[name].lastUsage = app.time();
 
-		return jsonMemory[name].json;
+		return app.copy(jsonMemory[name].json);
 	}
 
 	return false;